change to Long for version primitives (#79)

+++ app/src/main/java/cc/sovellus/vrcaa/api/vrchat/models/World.kt
@@ -59,7 +59,7 @@ data class World(
     @SerializedName("updated_at")
     var updatedAt: String = "",
     @SerializedName("version")
-    var version: Int = 0,
+    var version: Long = 0,
     @SerializedName("visits")
     var visits: Int = 0
 )
